requirements.txt
snap-standford==6.0.0
numpy
matplotlib
